# apps/api/src/routes/playlists.ts
import {Router} from 'express'; import {z} from 'zod'; import {q} from '../db.js'; import {auth,AuthRequest} from '../middleware/auth.js';
const r=Router();
const body=z.object({name:z.string().min(1).max(120),description:z.string().max(1000).optional(),visibility:z.enum(['PUBLIC','PRIVATE','UNLISTED']).default('PRIVATE'),collaborative:z.boolean().default(false)});
r.get('/',auth,async(req:AuthRequest,res,next)=>{try{const {rows}=await q<any>('SELECT * FROM playlists WHERE owner_id=$1 ORDER BY updated_at DESC',[req.user!.id]);res.json(rows)}catch(e){next(e)}});
r.post('/',auth,async(req:AuthRequest,res,next)=>{try{const x=body.parse(req.body);const {rows}=await q<any>('INSERT INTO playlists(owner_id,name,description,visibility,collaborative) VALUES($1,$2,$3,$4,$5) RETURNING *',[req.user!.id,x.name,x.description??'',x.visibility,x.collaborative]);res.status(201).json(rows[0]);}catch(e){next(e)}});
r.get('/:id',async(req,res,next)=>{try{const {rows}=await q<any>(`SELECT p.*,COALESCE(json_agg(json_build_object('id',s.id,'title',s.title,'artist_name',u.display_name,'artwork_url',s.artwork_url,'duration_ms',s.duration_ms) ORDER BY pt.position) FILTER (WHERE s.id IS NOT NULL),'[]') tracks FROM playlists p LEFT JOIN playlist_tracks pt ON pt.playlist_id=p.id LEFT JOIN songs s ON s.id=pt.song_id LEFT JOIN users u ON u.id=s.artist_id WHERE p.id=$1 GROUP BY p.id`,[req.params.id]);if(!rows[0])return res.status(404).json({error:'Playlist not found'});res.json(rows[0]);}catch(e){next(e)}});
r.post('/:id/tracks',auth,async(req:AuthRequest,res,next)=>{try{const x=z.object({songId:z.string().uuid()}).parse(req.body);const p=await q<any>('SELECT * FROM playlists WHERE id=$1',[req.params.id]);if(!p.rows[0])return res.status(404).json({error:'Playlist not found'});if(p.rows[0].owner_id!==req.user!.id&&!p.rows[0].collaborative)return res.status(403).json({error:'No edit permission'});const pos=await q<any>('SELECT COALESCE(MAX(position),-1)+1 position FROM playlist_tracks WHERE playlist_id=$1',[req.params.id]);await q('INSERT INTO playlist_tracks(playlist_id,song_id,position,added_by) VALUES($1,$2,$3,$4) ON CONFLICT DO NOTHING',[req.params.id,x.songId,pos.rows[0].position,req.user!.id]);await q('UPDATE playlists SET updated_at=now() WHERE id=$1',[req.params.id]);res.status(204).end();}catch(e){next(e)}}); export default r;