# apps/api/src/routes/subscriptions.ts
import {Router} from 'express'; import {auth,AuthRequest} from '../middleware/auth.js'; import {q} from '../db.js';
export interface BillingProvider{createSubscription(userId:string,plan:string):Promise<{externalId:string;status:'TRIALING'|'ACTIVE'}>;cancelSubscription(externalId:string):Promise<void>;restore(externalId:string):Promise<unknown>}
const r=Router(); r.get('/me',auth,async(req:AuthRequest,res,next)=>{try{const {rows}=await q<any>('SELECT * FROM subscriptions WHERE user_id=$1',[req.user!.id]);res.json(rows[0]??{status:'FREE',plan:'FREE'});}catch(e){next(e)}}); r.post('/checkout',auth,async(req:AuthRequest,res,next)=>{try{const plan=req.body.plan==='PREMIUM'?'PREMIUM':null;if(!plan)return res.status(400).json({error:'Unsupported plan'});const trial=new Date(Date.now()+30*86400000);const {rows}=await q<any>(`INSERT INTO subscriptions(user_id,provider,plan,status,trial_ends_at,current_period_end) VALUES($1,'PENDING','PREMIUM','TRIALING',$2,$2) ON CONFLICT(user_id) DO UPDATE SET plan='PREMIUM',status='TRIALING',trial_ends_at=$2,current_period_end=$2 RETURNING *`,[req.user!.id,trial]);res.status(201).json({subscription:rows[0],externalCheckoutRequired:true});}catch(e){next(e)}}); export default r;