# packages/contracts/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Waeve API
  version: 2.0.0
  description: Application-first global music platform API.
servers:
  - url: /api/v1
security: []
paths:
  /auth/register: {post: {summary: Register, requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/Register'}}}}, responses: {'201': {description: Created}}}}
...
  /me: {get: {summary: Current account, security: [{bearerAuth: []}]}, patch: {summary: Update account, security: [{bearerAuth: []}]}}
...
  /playlists: {get: {summary: List playlists, security: [{bearerAuth: []}]}, post: {summary: Create playlist, security: [{bearerAuth: []}]}}
  /playlists/{id}: {get: {summary: Playlist detail, security: [{bearerAuth: []}]}}
  /playlists/{id}/tracks: {post: {summary: Add track, security: [{bearerAuth: []}]}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  schemas:
    Register:
      type: object
      required: [email,password,displayName]
      properties:
        email: {type: string, format: email}
        password: {type: string, minLength: 12}
        displayName: {type: string, minLength: 2}
        countryCode: {type: string, minLength: 2, maxLength: 2}